IVC/folding: extract folding config
